+++ index.html
@@ -236,5 +236,7 @@
       </section>
     </template>
 
+    <script type="module" src="js/main.js"></script>
+    <script type="module" src="js/functions.js"></script>
 </body>
 </html>
